+++ doc/conf.py
@@ -100,7 +100,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Fedora Messaging'
-copyright = u'2012 Red Hat, Inc. and others.'
+copyright = u'2012 - 2014 Red Hat, Inc. and others.'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the