+++ fedmsg/utils.py
@@ -1,5 +1,5 @@
 # This file is part of fedmsg.
-# Copyright (C) 2012 Red Hat, Inc.
+# Copyright (C) 2012 - 2014 Red Hat, Inc.
 #
 # fedmsg is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -87,6 +87,35 @@ def set_tcp_keepalive(socket, config):
                 socket.setsockopt(attr, config[key])
 
 
+def set_tcp_reconnect(socket, config):
+    """ Set a series of TCP reconnect options on the socket if
+    and only if
+      1) they are specified explicitly in the config and
+      2) the version of pyzmq has been compiled with support
+
+    Once our fedmsg bus grew to include many hundreds of endpoints, we started
+    notices a *lot* of SYN-ACKs in the logs.  By default, if an endpoint is
+    unavailable, zeromq will attempt to reconnect every 100ms until it gets a
+    connection.  With this code, you can reconfigure that to back off
+    exponentially to some max delay (like 1000ms) to reduce reconnect storm
+    spam.
+
+    See the following
+      - http://api.zeromq.org/3-2:zmq-setsockopt
+    """
+
+    reconnect_options = {
+        # Map fedmsg config keys to zeromq socket constants
+        'zmq_reconnect_ivl': 'RECONNECT_IVL',
+        'zmq_reconnect_ivl_max': 'RECONNECT_IVL_max',
+    }
+    for key, const in reconnect_options.items():
+        if key in config:
+            attr = getattr(zmq, const, None)
+            if attr:
+                socket.setsockopt(attr, config[key])
+
+
 def load_class(location):
     """ Take a string of the form 'fedmsg.consumers.ircbot:IRCBotConsumer'
     and return the IRCBotConsumer class.
@@ -95,7 +124,7 @@ def load_class(location):
     tokens = mod_name.split('.')
 
     fromlist = '[]'
-    if tokens > 1:
+    if len(tokens) > 1:
         fromlist = '.'.join(tokens[:-1])
 
     module = __import__(mod_name, fromlist=fromlist)